feat(generate_records): implement game record generator with progress reporting
Added functionality to generate game records using the GameRecordGeneratorConfig.
Included a simple progress reporter sample implementation using the indicatif crate.
Training environment is still under development.

+++ temp_reversi_eval/src/feature_offsets.rs
@@ -1,6 +1,6 @@
 use super::patterns::PATTERNS;
 
-pub const fn make_feature_offsets() -> [u32; PATTERNS.len()] {
+pub const fn make_feature_offsets() -> [u16; PATTERNS.len()] {
     let mut offsets = [0; PATTERNS.len()];
     let mut offset = 0;
     let mut i = 0;
@@ -9,13 +9,13 @@ pub const fn make_feature_offsets() -> [u32; PATTERNS.len()] {
         offsets[i + 1] = offset;
         offsets[i + 2] = offset;
         offsets[i + 3] = offset;
-        offset = offset + PATTERNS[i].len() as u32;
+        offset = offset + PATTERNS[i].len() as u16;
         i += 4;
     }
     offsets
 }
 
-pub const FEATURE_OFFSETS: [u32; PATTERNS.len()] = make_feature_offsets();
+pub const FEATURE_OFFSETS: [u16; PATTERNS.len()] = make_feature_offsets();
 
 #[cfg(test)]
 mod tests {
@@ -28,9 +28,9 @@ mod tests {
         assert_eq!(offsets[1], 0);
         assert_eq!(offsets[2], 0);
         assert_eq!(offsets[3], 0);
-        assert_eq!(offsets[4], PATTERNS[4].len() as u32);
-        assert_eq!(offsets[5], PATTERNS[4].len() as u32);
-        assert_eq!(offsets[6], PATTERNS[4].len() as u32);
-        assert_eq!(offsets[7], PATTERNS[4].len() as u32);
+        assert_eq!(offsets[4], PATTERNS[4].len() as u16);
+        assert_eq!(offsets[5], PATTERNS[4].len() as u16);
+        assert_eq!(offsets[6], PATTERNS[4].len() as u16);
+        assert_eq!(offsets[7], PATTERNS[4].len() as u16);
     }
 }
